chore: refactor (gemini)
